Summarise session info at end of run
